requirements.txt
Flask==2.0.1
Flask-JSON==0.3.4
ollama-client==1.0.0
websocket-client
requests-toolbelt